Rename vc-constants finally
